chore: workspace clippy cleanup on cuda,no-cuda lib+tests
Mix of auto-fix and manual cleanup:
- main.rs: rewrite two match Ok/Err→return branches as let-else for
  clarity (auto-fix produced awkward if-let-else-return).
- ops/linear.rs + scheduler/cuda/prefill.rs: split doc-list paragraphs
  with a blank line so clippy's doc_lazy_continuation is satisfied.
- ops/tests.rs + carnice_dtype_audit.rs: hoist inner `use` / `struct`
  above let bindings (items-after-statements).
- infer-cli/hardware.rs, tests/*.rs: clippy --fix pass (dead import
  removed, needless `return`/`as i32`/`mut` dropped, filter_map closure
  replaced by `Result::ok`, clippy::map_unwrap_or conversions, etc.).
- seven tests: add `#[ignore = "..."]` reasons describing required
  weights/GPU.

CI (`--features no-cuda --lib`) still clean. cuda,no-cuda lib+tests
warnings dropped to mostly long-literal separator suggestions on Q4K
scale-factor constants and test code.

+++ infer/tests/q4k_kernel_correctness.rs
@@ -256,7 +256,7 @@ fn dequant_q3k_superblock_cpu(sb: &[u8; 110]) -> [f32; 256] {
     let mut out = [0f32; 256];
     for j in 0..256 {
         let q2 = (qs[j / 4] >> ((j % 4) * 2)) & 0x03;
-        let hbit = ((hmask[j / 8] >> (j % 8)) & 1) as u8;
+        let hbit = (hmask[j / 8] >> (j % 8)) & 1;
         let q3 = q2 | (hbit << 2);
         let sc = scales[j / 16] as f32;
         out[j] = d * sc * (q3 as f32 - 4.0);
